fix excessive wname values
In the pack/unpack code, we provide a limit (L9P_MAX_WELEM)
on the number of wname strings accepted in a Twalk.  If
the incoming value was too big, however, we would run off
the end of our own local array.

While here, convert all I/O operations to use ssize_t for
consistency, fix the return values of l9p_pustrings() and
l9p_puqids() (no one is checking them so this has no effect),
make l9p_pustring check for malloc failures, and mark
a slightly bogus assumption in l9p_pufcall().

Comment liberally.

+++ lib9p.h
@@ -145,7 +145,7 @@ struct l9p_backend {
 
 int l9p_pufcall(struct l9p_message *msg, union l9p_fcall *fcall,
     enum l9p_version version);
-int l9p_pustat(struct l9p_message *msg, struct l9p_stat *s,
+ssize_t l9p_pustat(struct l9p_message *msg, struct l9p_stat *s,
     enum l9p_version version);
 uint16_t l9p_sizeof_stat(struct l9p_stat *stat, enum l9p_version version);
 int l9p_pack_stat(struct l9p_message *msg, struct l9p_request *req,
